refactor(output): rename to report
