+++ runner/internal/report/report_test.go
@@ -1,4 +1,4 @@
-package output_test
+package report_test
 
 import (
 	"net/url"
@@ -9,7 +9,7 @@ import (
 
 	"github.com/benchttp/engine/runner/internal/config"
 	"github.com/benchttp/engine/runner/internal/metrics"
-	"github.com/benchttp/engine/runner/internal/output"
+	"github.com/benchttp/engine/runner/internal/report"
 )
 
 func TestReport_String(t *testing.T) {
@@ -18,7 +18,7 @@ func TestReport_String(t *testing.T) {
 	t.Run("return default summary if template is empty", func(t *testing.T) {
 		const tpl = ""
 
-		rep := output.New(newMetrics(), newConfigWithTemplate(tpl), d)
+		rep := report.New(newMetrics(), newConfigWithTemplate(tpl), d)
 		checkSummary(t, rep.String())
 	})
 
@@ -26,7 +26,7 @@ func TestReport_String(t *testing.T) {
 		const tpl = "{{ .Metrics.TotalCount }}"
 
 		m := newMetrics()
-		rep := output.New(m, newConfigWithTemplate(tpl), d)
+		rep := report.New(m, newConfigWithTemplate(tpl), d)
 
 		if got, exp := rep.String(), strconv.Itoa(m.TotalCount); got != exp {
 			t.Errorf("\nunexpected output\nexp %s\ngot %s", exp, got)
@@ -36,7 +36,7 @@ func TestReport_String(t *testing.T) {
 	t.Run("fallback to default summary if template is invalid", func(t *testing.T) {
 		const tpl = "{{ .Marcel.Patulacci }}"
 
-		rep := output.New(newMetrics(), newConfigWithTemplate(tpl), d)
+		rep := report.New(newMetrics(), newConfigWithTemplate(tpl), d)
 		got := rep.String()
 		split := strings.Split(got, "Falling back to default summary:\n")
