+++ runner/internal/report/template_internal_test.go
@@ -1,4 +1,4 @@
-package output
+package report
 
 import (
 	"errors"
